src/main/java/seedu/address/model/client/DisposableIncome.java
package seedu.address.model.client;

import static java.util.Objects.requireNonNull;
import static seedu.address.commons.util.AppUtil.checkArgument;
import static seedu.address.commons.util.StringUtil.getCurrencyFormat;
import static seedu.address.commons.util.StringUtil.isValidCurrencyValue;
import static seedu.address.commons.util.StringUtil.isWithinLengthLimit;

public class DisposableIncome extends NumberComparable<DisposableIncome>
    implements OptionalStringBasedField, ShorterFieldLength {

    public static final String MESSAGE_CONSTRAINTS =
            "Disposable Income numbers should be a non-negative number only. (Character limit: 15)";
    public static final String DEFAULT_VALUE = "0.00";

    public final String value;
    public final String valueWithSymbol;

    /**
     * Constructs a {@code Risk Appetite}.
     *
     * @param disposableIncome A valid risk appetite number.
     */
    public DisposableIncome(String disposableIncome) {
        requireNonNull(disposableIncome);
        checkArgument(isValidDisposableIncome(disposableIncome), MESSAGE_CONSTRAINTS);
        if (disposableIncome.isEmpty()) {
            disposableIncome = DEFAULT_VALUE;
        }
        valueWithSymbol = getCurrencyFormat(disposableIncome, true);
        value = getCurrencyFormat(disposableIncome, false);

    }

    /**
     * Returns true if a given string is a valid phone number.
     */
    public static boolean isValidDisposableIncome(String test) {
        return (IS_BLANK_VALUE_ALLOWED && test.isEmpty())
            || (isValidCurrencyValue(test) && isWithinLengthLimit(test, MAX_LENGTH));
    }

    @Override
    public int hashCode() {
        return value.hashCode();
    }

    @Override
    public boolean equals(Object other) {
        return other == this // short circuit if same object
            || (other instanceof DisposableIncome // instanceof handles nulls
            && value.equals(((DisposableIncome) other).value)); // state check
    }

    @Override
    public String toString() {
        return value;
    }
}
